requirements.txt
discord
pandas
requests
beautifulsoup4
lxml
pytz
mysql-connector-python